[rbrowsable] introduce ROOTObjectDrawProvider library for TObject
Implements generic methods for any TObject-derived classed for drawing
on TCanvas/RCanvas
