+++ gui/browserv7/src/RBrowser.cxx
@@ -533,66 +533,3 @@ void ROOT::Experimental::RBrowser::ProcessMsg(unsigned connid, const std::string
       ProcessSaveFile(arg.substr(9));
    }
 }
-
-
-// ============================================================================================
-
-using namespace ROOT::Experimental;
-
-/** Provider for drawing of ROOT6 classes */
-
-class RV6DrawProvider : public RDrawableProvider {
-public:
-
-   RV6DrawProvider()
-   {
-      RegisterV6(nullptr, [](TVirtualPad *pad, std::unique_ptr<Browsable::RHolder> &obj, const std::string &opt) -> bool {
-
-         // try take object without ownership
-         auto tobj = obj->get_object<TObject>();
-         if (!tobj) {
-            auto utobj = obj->get_unique<TObject>();
-            if (!utobj)
-               return false;
-            tobj = utobj.release();
-            tobj->SetBit(TObject::kMustCleanup); // TCanvas should care about cleanup
-         }
-
-         pad->GetListOfPrimitives()->Clear();
-
-         pad->GetListOfPrimitives()->Add(tobj, opt.c_str());
-
-         return true;
-      });
-   }
-
-} newRV6DrawProvider;
-
-
-/** Provider for drawing of ROOT7 classes */
-
-class RV7DrawProvider : public RDrawableProvider {
-public:
-   RV7DrawProvider()
-   {
-      RegisterV7(nullptr, [] (std::shared_ptr<RPadBase> &subpad, std::unique_ptr<Browsable::RHolder> &obj, const std::string &opt) -> bool {
-
-         // here clear ownership is required
-         // If it possible, TObject will be cloned by RTObjectHolder
-         auto tobj = obj->get_shared<TObject>();
-         if (!tobj) return false;
-
-         if (subpad->NumPrimitives() > 0) {
-            subpad->Wipe();
-            subpad->GetCanvas()->Modified();
-            subpad->GetCanvas()->Update(true);
-         }
-
-         subpad->Draw<RObjectDrawable>(tobj, opt);
-         return true;
-      });
-
-   }
-
-} newRV7DrawProvider;
-
